+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Fixed
+
+- Remove empty expanded sdkconfig files folder after build
+- Split up expanded sdkconfig files folder for different build
+
 ## [0.5.1] (2023-04-06)
 
 ### Fixed